Enhance README with architecture diagram, setup instructions, and evaluation criteria

+++ product-service/src/main/java/com/example/product/dto/ProductResponse.java
@@ -19,5 +19,5 @@ public class ProductResponse {
     private String nombre;
 
     @Schema(description = "Precio del producto", example = "999.99", type = "number", format = "double")
-    private Double precio;
+    private BigDecimal precio;
 }
